Decouple HTTP handlers
