+++ backend/daemon/daemon.go
@@ -4,20 +4,16 @@ package daemon
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"net"
 	"net/http"
-	"runtime/debug"
 	"strconv"
-	"strings"
 	"time"
 
 	"mintter/backend/config"
 	"mintter/backend/core"
 	"mintter/backend/daemon/api"
 	"mintter/backend/daemon/storage"
-	"mintter/backend/graphql"
 	"mintter/backend/hyper"
 	"mintter/backend/ipfs"
 	"mintter/backend/logging"
@@ -28,9 +24,7 @@ import (
 	"mintter/backend/wallet"
 
 	"crawshaw.io/sqlite/sqlitex"
-	"github.com/99designs/gqlgen/graphql/playground"
 	"github.com/gorilla/mux"
-	"github.com/improbable-eng/grpc-web/go/grpcweb"
 	"github.com/libp2p/go-libp2p/core/crypto"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/collectors"
@@ -39,7 +33,6 @@ import (
 	"go.opentelemetry.io/otel/sdk/trace"
 	"go.uber.org/multierr"
 	"go.uber.org/zap"
-	"golang.org/x/exp/slices"
 	"golang.org/x/sync/errgroup"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/reflection"
@@ -174,7 +167,7 @@ func LoadWithStorage(ctx context.Context, cfg config.Config, r *storage.Dir, grp
 
 		return fileManager.Start(n.Blobs().IPFSBlockstore(), n.Bitswap(), n.Provider())
 	})
-	a.HTTPServer, a.HTTPListener, err = initHTTP(cfg.HTTP.Port, a.GRPCServer, &a.clean, a.g, a.DB, a.Net, me, a.Wallet, fileManager)
+	a.HTTPServer, a.HTTPListener, err = initHTTP(cfg.HTTP.Port, a.GRPCServer, &a.clean, a.g, a.Wallet, fileManager)
 	if err != nil {
 		return nil, err
 	}
@@ -457,12 +450,13 @@ const (
 	routeNav    = 1 << 2
 )
 
-type router struct {
+// Router is a wrapper around mux that can build the navigation menu.
+type Router struct {
 	r   *mux.Router
 	nav []string
 }
 
-func (r *router) Handle(path string, h http.Handler, mode int) {
+func (r *Router) Handle(path string, h http.Handler, mode int) {
 	h = instrumentHTTPHandler(h, path)
 
 	if mode&routePrefix != 0 {
@@ -476,137 +470,13 @@ func (r *router) Handle(path string, h http.Handler, mode int) {
 	}
 }
 
-func (r *router) Index(w http.ResponseWriter, req *http.Request) {
+func (r *Router) Index(w http.ResponseWriter, req *http.Request) {
 	for _, route := range r.nav {
 		fmt.Fprintf(w, `<p><a href="%s">%s</a></p>`, route, route)
 	}
 }
 
-func initHTTP(
-	port int,
-	rpc *grpc.Server,
-	clean *cleanup.Stack,
-	g *errgroup.Group,
-	db *sqlitex.Pool,
-	node *future.ReadOnly[*mttnet.Node],
-	me *future.ReadOnly[core.Identity],
-	wallet *wallet.Service,
-	ipfsHandler ipfs.HTTPHandler,
-) (srv *http.Server, lis net.Listener, err error) {
-	var h http.Handler
-	{
-		grpcWebHandler := grpcweb.WrapServer(rpc, grpcweb.WithOriginFunc(func(origin string) bool {
-			return true
-		}))
-
-		router := router{r: mux.NewRouter()}
-		router.Handle("/debug/metrics", promhttp.Handler(), routeNav)
-		router.Handle("/debug/pprof", http.DefaultServeMux, routePrefix|routeNav)
-		router.Handle("/debug/vars", http.DefaultServeMux, routePrefix|routeNav)
-		router.Handle("/debug/grpc", grpcLogsHandler(), routeNav)
-		router.Handle("/debug/buildinfo", buildInfoHandler(), routeNav)
-		router.Handle("/graphql", corsMiddleware(graphql.Handler(wallet)), 0)
-		router.Handle("/playground", playground.Handler("GraphQL Playground", "/graphql"), routeNav)
-		router.Handle(ipfs.IPFSRootRoute+ipfs.UploadRoute, http.HandlerFunc(ipfsHandler.UploadFile), 0)
-		router.Handle(ipfs.IPFSRootRoute+ipfs.GetRoute, http.HandlerFunc(ipfsHandler.GetFile), 0)
-
-		router.r.MatcherFunc(mux.MatcherFunc(func(r *http.Request, match *mux.RouteMatch) bool {
-			return grpcWebHandler.IsAcceptableGrpcCorsRequest(r) || grpcWebHandler.IsGrpcWebRequest(r)
-		})).Handler(grpcWebHandler)
-
-		router.Handle("/", http.HandlerFunc(router.Index), 0)
-
-		h = router.r
-	}
-
-	srv = &http.Server{
-		Addr:              ":" + strconv.Itoa(port),
-		ReadHeaderTimeout: 5 * time.Second,
-		// WriteTimeout:      10 * time.Second,
-		IdleTimeout: 20 * time.Second,
-		Handler:     h,
-	}
-
-	lis, err = net.Listen("tcp", srv.Addr)
-	if err != nil {
-		return
-	}
-
-	g.Go(func() error {
-		err := srv.Serve(lis)
-		if err == http.ErrServerClosed {
-			return nil
-		}
-		return err
-	})
-
-	clean.AddErrFunc(func() error {
-		return srv.Shutdown(context.Background())
-	})
-
-	return
-}
-
-// corsMiddleware allows different host/origins.
-func corsMiddleware(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		// allow cross domain AJAX requests
-		w.Header().Set("Access-Control-Allow-Origin", "*")
-		w.Header().Set("Access-Control-Allow-Headers", "Origin, X-Requested-With, Content-Type, Accept")
-		next.ServeHTTP(w, r)
-	})
-}
-
 // WithMiddleware generates an grpc option with the given middleware.
 func WithMiddleware(i grpc.UnaryServerInterceptor) grpc.ServerOption {
 	return grpc.UnaryInterceptor(i)
-}
-
-// GwEssentials is a middleware to restrict incoming grpc calls to bare minimum for the gateway to work.
-func GwEssentials(ctx context.Context,
-	req interface{},
-	info *grpc.UnaryServerInfo,
-	handler grpc.UnaryHandler) (interface{}, error) {
-	methodSplitted := strings.Split(info.FullMethod, "/")
-	if len(methodSplitted) < 2 || (strings.ToLower(methodSplitted[len(methodSplitted)-1]) != "getpublication" &&
-		strings.ToLower(methodSplitted[len(methodSplitted)-1]) != "listcitations" &&
-		strings.ToLower(methodSplitted[len(methodSplitted)-1]) != "getaccount") {
-		return nil, fmt.Errorf("method: %s not allowed", info.FullMethod)
-	}
-
-	// Calls the handler
-	h, err := handler(ctx, req)
-
-	return h, err
-}
-
-func buildInfoHandler() http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		info, ok := debug.ReadBuildInfo()
-		if !ok {
-			http.Error(w, "doesn't support build info", http.StatusExpectationFailed)
-			return
-		}
-
-		// Don't want to show information about all the dependencies.
-		info.Deps = nil
-
-		// Want to support text and json.
-		wantJSON := slices.Contains(r.Header.Values("Accept"), "application/json") ||
-			r.URL.Query().Get("format") == "json"
-
-		if wantJSON {
-			w.Header().Set("Content-Type", "application/json")
-
-			enc := json.NewEncoder(w)
-			enc.SetIndent("", "  ")
-
-			if err := enc.Encode(info); err != nil {
-				http.Error(w, err.Error(), http.StatusInternalServerError)
-			}
-		} else {
-			w.Header().Set("Content-Type", "text/plain")
-			fmt.Fprint(w, info.String())
-		}
-	})
 }
